// src/app/shared/enums/apiUrl.ts
export enum ApiUrl{
  login = 'auth/login',
  customerSignup = 'auth/register/user',
  driverSignup = 'auth/register/driver',
  factorySignup = 'auth/register/factory',
  companySignup = 'auth/register/drivercompany',
  registerDriverInfo = 'users/driverinfo',
  sendOtp = 'auth/sendOtp',
  verifyOtp = 'auth/verifyOtp',
  uploadDriverImages = 'users/driver/image',
  editCustomerDetails = 'users/normal',
  editFactoryDetails = 'users/factory',
  editDriverDetails = 'users/driver',
  editCompanyDetails = 'users/drivercompany',
  sendPasswordOtp = 'auth/password/otp',
  verifyPasswordOtp = 'auth/password/checkCode',
  resetPassword = 'auth/password/reset',
  uploadCustomerImage = 'users/normal/image',
  uploadFactoryImages = 'users/factory/image',
  uploadCompanyImages = 'users/drivercompany/image',
  supportMessage = 'supportCenter',
  getAllCareers = 'careers/index',
  getAllArticles = 'articles/index',
  getArticleById = 'articles/show',
  ordersByUserId = 'orders/getByUserId',
}